# perf: analyze cjs exports and emit typescript in parallel