--- a/ext/node/Cargo.toml
+++ b/ext/node/Cargo.toml
@@ -16,6 +16,7 @@ path = "lib.rs"
 [dependencies]
 aead-gcm-stream = "0.1"
 aes.workspace = true
+async-trait.workspace = true
 brotli.workspace = true
 bytes.workspace = true
 cbc.workspace = true