… Cargo.lock
… cli/resolver.rs
… ext/node/analyze.rs
… tests/integration/inspector_tests.rs
… tests/specs/npm/local_dir_no_duplicate_resolution/main.out
… tests/specs/npm/local_dir_no_duplicate_resolution/run_main_sorted_lines.ts