Two main things we can do in parallel:

1. CJS export analysis
1. TypeScript emit

This makes the loader async to make doing this easier.

Deleting the node analysis cache between runs and importing some npm packages (this only makes the first run much faster due to the caches... the second run is the same speed as befroe):

```
Benchmark 1: ../deno/deno_old run -A ts_morph.ts
  Time (mean ± σ):     406.1 ms ±   6.5 ms    [User: 358.1 ms, System: 62.2 ms]
  Range (min … max):   395.9 ms … 412.9 ms    10 runs
 
Benchmark 2: ../deno/target/release/deno run -A ts_morph.ts
  Time (mean ± σ):     278.2 ms ±   9.3 ms    [User: 411.1 ms, System: 81.4 ms]
  Range (min … max):   266.1 ms … 292.9 ms    10 runs
 
Summary
  ../deno/target/release/deno run -A ts_morph.ts ran
    1.46 ± 0.05 times faster than ../deno/deno_old run -A ts_morph.ts
% cat ts_morph.ts
import "npm:ts-morph@22.0.0";
import "npm:ts-morph@21.0.0";
import "npm:ts-morph@20.0.0";
import "npm:ts-morph@19.0.0";
```

Deleting the TypeScript emit and module analysis cache between runs and importing multiple copies of Dax:

```
Benchmark 1: ../deno/deno_old run -A dax_deno_land_x.ts
  Time (mean ± σ):     587.3 ms ±  24.0 ms    [User: 347.5 ms, System: 221.0 ms]
  Range (min … max):   569.2 ms … 650.7 ms    10 runs
 
Benchmark 2: ../deno/target/release/deno run -A dax_deno_land_x.ts
  Time (mean ± σ):     489.3 ms ±  14.3 ms    [User: 554.8 ms, System: 411.2 ms]
  Range (min … max):   473.9 ms … 510.2 ms    10 runs
 
Summary
  ../deno/target/release/deno run -A dax_deno_land_x.ts ran
    1.20 ± 0.06 times faster than ../deno/deno_old run -A dax_deno_land_x.ts
scratch % cat dax_deno_land_x.ts
import "https://deno.land/x/dax@0.39.2/mod.ts";
import "https://deno.land/x/dax@0.38.0/mod.ts";
import "https://deno.land/x/dax@0.37.0/mod.ts";
import "https://deno.land/x/dax@0.36.0/mod.ts";
import "https://deno.land/x/dax@0.35.0/mod.ts";
```

Note that this is still parsing typescript files sequentially because it doesn't include (https://github.com/denoland/deno_graph/pull/477). It is just emitting typescript files in parallel.